LayoutTests/platform/ios-sim-deprecated/media/media-source/media-source-duplicate-seeked-expected.txt

CONSOLE MESSAGE: line 19: TypeError: undefined is not a constructor (evaluating 'new MediaSource()')
FAIL: Timed out waiting for notifyDone to be called
